Updated cascading for destinations and trips. Fixing issues with implementation and testing for deleting destinations.
#story[456]

+++ app/controllers/destinations/DestinationController.java
@@ -118,7 +118,12 @@ public class DestinationController extends Controller {
      */
     public Result fetch(Http.Request request) {
 
-        Profile loggedInUser = profileRepo.fetchSingleProfile(AuthenticationUtil.getLoggedInUserId(request));
+        Integer loggedInUserId = AuthenticationUtil.getLoggedInUserId(request);
+        if (loggedInUserId == null) {
+            return unauthorized();
+        }
+
+        Profile loggedInUser = profileRepo.fetchSingleProfile(loggedInUserId);
 
         int pageNumber = 0;
         int pageSize = 50;
@@ -360,14 +365,24 @@ public class DestinationController extends Controller {
     /**
      * Deletes a destination from the database using the given destination id number.
      *
-     * @param id    the id of the destination.
-     * @return      notFound() (Http 404) if destination could not found, ok() (Http 200) if successfully deleted.
+     * @param destinationId     the id of the destination.
+     * @return                  notFound() (Http 404) if destination could not found, ok() (Http 200) if successfully deleted.
      */
-    public Result destroy(Long id) {
-        Destination destination = Destination.find.byId(id.intValue());
+    public Result destroy(Http.Request request, Long destinationId) {
+        Integer loggedInUserId = AuthenticationUtil.getLoggedInUserId(request);
+        if (loggedInUserId == null) {
+            return unauthorized();
+        }
+
+        Destination destination = Destination.find.byId(destinationId.intValue());
 
         if (destination == null) {
             return notFound();
+        }
+
+        Profile loggedInUser = profileRepo.fetchSingleProfile(loggedInUserId);
+        if (!AuthenticationUtil.validUser(loggedInUser, destination.getOwner())) {
+            return forbidden();
         }
 
         destinationRepo.delete(destination);
